Setup DI, Retrofit with basic test implementation - moving Interceptor
